Fix version comparisions in Facebook::Opengraph::Response
When using Facebook API version 3.2, the comparision in
Facebook::OpenGraph->_get_token that tests if the response is 2.3 is later
failed.

This happens because is_api_version_eq_or_later_than compares the API versions
like this:

$comp_major <= $response_major && $comp_minor <= $response_minor;

However while $comp_major is less than $response_major, the same does not hold
for the minor versions. The minor version only needs to be less than, iff the
major versions are the same.

This patch fixes this issue in both is_api_version_eq_or_later_than and
is_api_version_eq_or_older_than, and adds tests to verify that it works as
expected.